# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food Delivery API",
        "description": "API documentation for the food delivery system.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/kin-info-update": {
            "post": {
                "tags": ["Profile"],
                "summary": "Update next of kin information",
                "operationId": "updateKinInfo",
                "security": [
                { "bearerAuth": [] }
                ],
                "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                    "schema": {
                        "type": "object",
                        "required": ["name", "phone", "relationship", "address"],
                        "properties": {
                        "name": { "type": "string" },
                        "phone": { "type": "string" },
                        "relationship": { "type": "string" },
                        "address": { "type": "string" }
                        }
                    }
                    }
                }
                },
                "responses": {
                "200": {
                    "description": "Kin Info Updated Successful",
                    "content": {
                    "application/json": {
                        "schema": {
                        "type": "object",
                        "properties": {
                            "code": { "type": "integer", "example": 200 },
                            "success": { "type": "boolean", "example": true },
                            "message": { "type": "string", "example": "Kin Info Updated Successful" }
                        }
                        }
                    }
                    }
                }
                }
            }
        },
        "/bank-info-update": {
            "post": {
                "tags": ["Profile"],
                "summary": "Update bank account information",
                "operationId": "updateBankAccInfo",
                "security": [
                { "bearerAuth": [] }
                ],
                "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                    "schema": {
                        "type": "object",
                        "required": ["name", "acc_number", "sort_code", "is_account_in_your_name", "name_on_the_account"],
                        "properties": {
                        "name": { "type": "string" },
                        "acc_number": { "type": "string" },
                        "sort_code": { "type": "string" },
                        "is_account_in_your_name": { "type": "boolean" },
                        "name_on_the_account": { "type": "string" }
                        }
                    }
                    }
                }
                },
                "responses": {
                "200": {
                    "description": "Bank Account Info Updated Successful",
                    "content": {
                    "application/json": {
                        "schema": {
                        "type": "object",
                        "properties": {
                            "code": { "type": "integer", "example": 200 },
                            "success": { "type": "boolean", "example": true },
                            "message": { "type": "string", "example": "Bank Account Info Updated Successful" }
                        }
                        }
                    }
                    }
                }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}